refactor(ai): rename chat-stream-wrapper to chat-stream-summarize
