+++ packages/typescript/ai/src/activities/summarize/index.ts
@@ -316,4 +316,4 @@ export { BaseSummarizeAdapter } from './adapter'
 export {
   ChatStreamSummarizeAdapter,
   type ChatStreamCapable,
-} from './chat-stream-wrapper'
+} from './chat-stream-summarize'
